requirements.txt
openai>=1.14.0
python-dotenv>=1.0.0
pydantic>=2.0.0
aiohttp>=3.8.0
openai-agent-toolkit>=0.1.11  # OpenAI Agent SDK